extract hash() method to ease re-use
